requirements.txt
Pillow
streamlit=
torch
torchvision